YUK-8318 - IVVC: Create a method for submitting a group of requests and grouping the responses.
Added GroupPointDataRequest.cpp

+++ yukon-server/CapControl/cccapbank.cpp
@@ -1987,7 +1987,7 @@ CtiCCCapBank& CtiCCCapBank::setTagsControlStatus(LONG tags)
 }
 
 
-CtiCCCapBank& CtiCCCapBank::addAllCapBankPointsToMsg(std::list<int>& pointAddMsg)
+CtiCCCapBank& CtiCCCapBank::addAllCapBankPointsToMsg(std::list<long>& pointAddMsg)
 {
 
     if( getStatusPointId() > 0 )
